% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add numbers together}
\usage{
add(x, y)
}
\arguments{
\item{x}{First number}

\item{y}{Second number}
}
\value{
Sum of the first and second number
}
\description{
Add numbers together
}
\examples{
add(1, 1)
add(20, 22)
}
